Fix: Render text after last annotation
When rendering annotated text, the portion of the text that came after the final annotation was being dropped.

This is fixed by adding a check to see if the current facet is the last one in the list. If it is, the remaining text from the end of the facet to the end of the content is appended.