SkeetView: Allow navigating to quoted skeet's thread
Previously, tapping on a quoted skeet within another skeet would navigate to the thread of the top-level skeet. This change correctly passes the `onShowThread` handler down through nested and embedded skeets.

Now, tapping on a quoted skeet will navigate to the thread for that specific skeet, as intended. This also fixes an issue where tapping a "see more" link on a quoted skeet would incorrectly show the parent's thread.